fix silly mistakes

+++ src/simple-cache/query_data.cc
@@ -200,7 +200,7 @@ namespace virtdb { namespace simple_cache {
              V_(qtl.t0_nblocks()) <<
              V_(t0) << V_(t1) <<
              V_(ret) << V_(res) <<
-             V_((int)difftime) <<
+             V_(difftime) <<
              V_(timeout_seconds_.load())
              );
